Added ( (Int32, Int32), Float) constructor

+++ Sources/RaylibSwift/Shape/Circle.swift
@@ -22,10 +22,10 @@ extension Circle {
         self.init(center, radius)
     }
 
-    /// Creates a circle with a Float tuple center point and radius
+    /// Creates a circle with a Int32 tuple center point and radius
     @inlinable
-    public init(center: (x: Float, y: Float), radius: Float) {
-        self.init(Vector2(center.x, center.y), radius)
+    public init(center: (x: Int32, y: Int32), radius: Float) {
+        self.init(Vector2(Float(center.x), Float(center.y)), radius)
     }
 
     /// Creates a circle with separate x/y coordinates and radius
